Fix memory corruption bug.
There are two problems, some of the init code needs the map created before
the init function since it does a verify check during init. The second
problem is that the local addr space seems to change for reasons I don't
understand, so it's necessary to init the map list when doing a unw_cursor
init.

Bug: 12852152

Change-Id: I32246d7c5adf7e24192868cb32a8c2fdc53c12f0