Use -Wno-header-guard for clang 3.5
Clang 3.5 misdetects the #ifndef UNW_REMOTE_ONLY/#define UNW_LOCAL_ONLY
construct in src/mi/backtrace.c as a header guard with mismatching macro
names, causing a warning in code built with -Werror by default.

While this is a false positive, there's really not much clang can do to
detect this (the code does look exactly like a header guard), so this is
not really a workaround for a compiler bug.

Change-Id: If1e9d8b4344719b706f6e1856c97dfcc2edaeb72
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>