Fix local arm unwind to handle dwarf properly.
It turns out that the local unwind for arm only checks the debug_frame
for dwarf information. This means that any arm assembly code that only
has dwarf information cannot be unwound. Fix this so the arm function
calls the dwarf_find_proc_info instead of doing the half implementation
that existed originally.

Bug: 17404548
Change-Id: Iffae860f2357d114fc004daa215e9d96a461379d